remove token_maps lib and moved maps to tokens
