+++ Cargo.toml
@@ -16,8 +16,9 @@ syntax = { path = "./crates/syntax" }
 parser = { path = "./crates/parser" }
 ast = { path = "./crates/ast" }
 tokens = { path = "./libs/tokens" }
-token_maps = { path = "./libs/token_maps" }
 macros = { path = "./libs/macros" }
 classpath_resolver = { path = "./libs/classpath_resolver" }
 pretty_assertions = "1.4.0"
 itertools = "0.12.1"
+assert-panic = "1.0.1"
+rstest = "0.19.0"
