+++ crates/lexer/Cargo.toml
@@ -10,10 +10,9 @@ rust-version.workspace = true
 
 [dependencies]
 tokens = { workspace = true }
-token_maps = { workspace = true }
 unicode_categories = "0.1.1"
 
 [dev-dependencies]
 macros = { workspace = true }
-rstest = "0.19.0"
+rstest.workspace = true
 pretty_assertions.workspace = true
