+++ crates/parser/Cargo.toml
@@ -18,5 +18,5 @@ lexer.workspace = true
 itertools.workspace = true
 ast.workspace = true
 rowan = "0.15.15"
-rstest = "0.19.0"
+rstest.workspace = true
 pretty_assertions.workspace = true
