+++ crates/syntax/Cargo.toml
@@ -15,5 +15,5 @@ lexer.workspace = true
 itertools.workspace = true
 
 [dev-dependencies]
-rstest = "0.19.0"
+rstest.workspace = true
 pretty_assertions.workspace = true
