--- libs/token_maps/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "token_maps"
-version = "0.0.1"
-description = "TBD"
-
-authors.workspace = true
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-phf = { version = "0.11", features = ["macros"] }
-tokens = { workspace = true }
